requirements.txt
fastapi
uvicorn
pydantic
pyyaml
requests
tqdm
psutil
cryptography
httpx
numpy
opencv-python-headless
Pillow
pytest
ruff
python-multipart
